+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 -----------------
 
 - Simulation wizard: Improve laterals page.
+- Improvements for Upload wizard (#541).
+- Updated threedi-api-client version to 4.1.6.
 
 3.10 (2024-04-12)
 -----------------